Add file assertions

+++ src/Infrastructure/Filesystem/Directory.php
@@ -11,6 +11,9 @@ class Directory
      */
     public function clear(): void
     {
+        $this->assertTrue($this->exists(), "Cannot clear {$this->path}: Directory does not exist");
+        $this->assertTrue($this->isWritable(), "Cannot clear {$this->path}: Directory is not writable");
+
         foreach ($this->list() as $item) {
             if ($item instanceof Directory) {
                 $item->clear();
@@ -24,6 +27,8 @@ class Directory
      */
     public function create(): void
     {
+        $this->assertTrue(!$this->exists(), "Cannot create {$this->path}: File or directory already exists");
+
         mkdir($this->path);
     }
 
@@ -32,7 +37,11 @@ class Directory
      */
     public function delete(): void
     {
-        rmdir($this->path);
+        if ($this->exists()) {
+            $this->assertTrue($this->isEmpty(), "Cannot delete {$this->path}: Directory is not empty");
+            $this->assertTrue($this->isWritable(), "Cannot delete {$this->path}: Directory is not writable");
+            rmdir($this->path);
+        }
     }
 
     /**
@@ -52,6 +61,9 @@ class Directory
      */
     public function list(): array
     {
+        $this->assertTrue($this->exists(), "Cannot list {$this->path}: Directory does not exist");
+        $this->assertTrue($this->isReadable(), "Cannot list {$this->path}: Directory is not readable");
+
         $items = [];
         $handle = opendir($this->path);
         while (false !== ($item = readdir($handle))) {
